Add dispatcher shied and remove duplicated services
Change of strategy. The double resource/file/auth services was a good
idea and was working but not perfectly. The session roles were mixed-up
when switching from logged-in to regular user, causing for example
logged in users not to be able to view the complaints, or regular users
not to be able to submit a complaint if they were logged in right before
that.

This new approach keeps things actually simpler, we keep the data in the
public graph as it used to be and have only one set of services, but we
add a shield in the dispatcher that checks in the headers if we do have
the expected role when trying to GET the complaints or the files.